+++ pkg/domain/entities/types.go
@@ -2,6 +2,7 @@ package entities
 
 import (
 	"net"
+	"os"
 
 	buildahDefine "github.com/containers/buildah/define"
 	"github.com/containers/common/libnetwork/types"
@@ -110,6 +111,11 @@ type ContainerCreateResponse struct {
 // BuildOptions describe the options for building container images.
 type BuildOptions struct {
 	buildahDefine.BuildOptions
+	ContainerFiles []string
+	// Files that need to be closed after the build
+	// so need to pass this to the main build functions
+	LogFileToClose *os.File
+	TmpDirToClose  string
 }
 
 // BuildReport is the image-build report.
